requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==0.5.0
django==5.2.4
django-cors-headers==4.7.0
django-cors-middleware==1.5.0
djangorestframework==3.15.2
dotenv==0.9.9
gunicorn==23.0.0
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.7.0